Move Blob
