+++ src/util/blob.ts
@@ -1,0 +1,8 @@
+export interface Blob {
+  readonly size: number;
+  readonly type: string;
+  arrayBuffer(): Promise<ArrayBuffer>;
+  slice(start?: number, end?: number, contentType?: string): Blob;
+  stream(): any;
+  text(): Promise<string>;
+}
